requirements.txt
apache-airflow==2.8.1
pyspark==3.5.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
requests==2.28.1